requirements.txt
discord.py==2.4.0
yt-dlp==2024.8.6
PyNaCl==1.5.0
asyncio==3.4.3
opuslib==3.0.1
python-dotenv==1.0.1